Rebase master ok, adding functions to create a board, check if valid, tests on check and semi architecture
