+++ src/args/mod.rs
@@ -1,1 +1,4 @@
-mod parser;
+// pub use self::parser::load_file;
+// mod parser;
+
+pub mod parser;
