+++ src/board/mod.rs
@@ -1,2 +1,3 @@
 pub mod check;
 pub mod create;
+pub mod utils;
