+++ src/lib.rs
@@ -1,1 +1,3 @@
 pub mod board;
+pub mod args;
+pub mod algo;
